Fix CI test based on the requirements of the new merlin loader  (#536)
* cast targets to int64 dtype required by torch.scatter_

* update the dataset schema to use the same as the one defined in the model

* update exomrees46 data used in ci with hex feature converted to numerical

+++ transformers4rec/torch/losses.py
@@ -31,7 +31,7 @@ class LabelSmoothCrossEntropyLoss(_WeightedLoss):
             targets = (
                 torch.empty(size=(targets.size(0), n_classes), device=targets.device)
                 .fill_(smoothing / (n_classes - 1))
-                .scatter_(1, targets.data.unsqueeze(1), 1.0 - smoothing)
+                .scatter_(1, targets.data.unsqueeze(1).to(torch.int64), 1.0 - smoothing)
             )
         return targets
